requirements.txt
streamlit==1.31.0
easyocr==1.7.1
opencv-python==4.9.0.80
pymongo==4.6.1
pillow==10.2.0
numpy==1.26.3
dnspython==2.5.0
requests==2.31.0
openfoodfacts>=3.0.0
